requirements.txt
bcrypt
pycryptodome
cryptography
requests
colorama